% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_compare.R
\name{baseline_compare}
\alias{baseline_compare}
\title{Compare values to a baseline}
\usage{
baseline_compare(
  val_df,
  group_col,
  base_lev,
  values_col,
  comp_fun,
  ...,
  failmissing = TRUE,
  names_to = "name",
  values_to = "value"
)
}
\arguments{
\item{val_df}{dataframe of values}

\item{group_col}{column or columns to use for grouping- needed if the baseline is one level in this column. Otherwise ignored}

\item{base_lev}{value for baseline. Can be a level in \code{group_col}, or a scalar, or a vector of \code{length(unique(val_df$group_col))}. A dataframe would be useful but not yet implemented.}

\item{values_col}{column containing the values. Can be character or tidyselect or bare names, though the last are fragile.}

\item{comp_fun}{function to use for comparison comparing the first two arguments (\code{difference} and \code{relative} included here as wrappers for \code{-} and \code{/}). Can be character, or list-format, (both safe) or bare name (brittle)}

\item{...}{additional arguments to \code{\link[=create_base]{create_base()}}}

\item{failmissing}{logical, default TRUE. Use \code{tidyselect::any_of} or \code{tidyselect::all_of} when \code{values_col} or \code{group_col} are character. see \code{\link[=selectcreator]{selectcreator()}}}

\item{names_to}{character, as in \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}, used for auto-pivotting}

\item{values_to}{character, as in \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}, used for auto-pivotting}
}
\value{
a tibble matching \code{val_df} (or a long version thereof if \code{val_df} is wide), with an added column for the reference level (named \code{ref_values_col}) and a column of the compared values (named \code{comp_fun_values_col})
}
\description{
This is used to automate comparison to baselines of various sorts without needing to do all the data wrangling by hand.
}
