% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ewr_output.R
\name{cleanYrEWRs}
\alias{cleanYrEWRs}
\title{Clean the OLD STYLE incoming annual EWR results}
\usage{
cleanYrEWRs(ewrdf)
}
\arguments{
\item{ewrdf}{incoming tibble of EWRs after read-in}
}
\value{
tibble reformatted and cleaned for ongoing analysis
}
\description{
\emph{This function is deprecated- the new annual EWR outputs are much cleaner on
read-in. This still exists to allow use of historic versions.} There is some
data organisation that happens for the toolkit in the scenario controller,
but this does a bit more cleaning of data format for further analysis
}
