% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_agg_funs.R
\name{CompensatingFactor}
\alias{CompensatingFactor}
\title{Compensating aggregation}
\usage{
CompensatingFactor(x, na.rm = TRUE)
}
\arguments{
\item{x}{values to aggregate}

\item{na.rm}{default \code{na.rm = TRUE} finds the aggregation ignoring NA, opposite usual na.rm default \code{na.rm = FALSE}, which returns NA if there are any NA.}
}
\value{
summarised data
}
\description{
Doing well in one time/place can compensate for doing poorly elsewhere. Implemented with max(). For binary, this means passing once means passing the aggregated value.
}
