% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_comparison_funs.R
\name{difference}
\alias{difference}
\title{Difference}
\usage{
difference(x, y)
}
\arguments{
\item{x}{data to compare}

\item{y}{reference data}
}
\description{
Easily-callable wrapper for \code{-}
}
