% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vals_causal.R
\name{extract_vals_causal}
\alias{extract_vals_causal}
\title{Extracts and arranges the aggregated values for use in causal network plots.}
\usage{
extract_vals_causal(agglist, whichaggs, valcol, targetlevels = names(agglist))
}
\arguments{
\item{agglist}{a list of tibbles of each step in the theme aggregation
(output of \code{theme_agg_multi} with \code{saveintermediate = TRUE}). Currently
assumes the aggregation history is already in columns (either
\code{theme_agg_multi} with \code{namehistory = FALSE} or have run
\code{agg_names_to_cols}). Would be fairly easy to detect and add a call to
\code{agg_names_to_cols} if this becomes an issue.}

\item{whichaggs}{a character vector of the desired sequence of aggregation
functions. This has to have only values used in the aggregation used to
create \code{agglist}, and can have only one entry per step (unlike in the
creation of \code{agglist}). If different aggregation histories are required,
this should be run again, because it is not appropriate to mix histories in
a causal plot.}

\item{valcol}{character, name of the column with the aggregated values.}

\item{targetlevels}{character, names of levels to include. Defaults to \code{names(agglist)}. Any subset of those names can be used.}
}
\value{
a tibble with grouping variables for scenario and gauge, a column for
values named the same as \code{valcol}, and \code{Name} and \code{NodeType} columns to
allow joining to the node df.
}
\description{
This takes the aggregations at each stage along the theme axis,
extracts the aggregated values for a given aggregation history (ie the
sequence of aggregation functions), and returns in a form that can be added
to edges or nodes dfs.
}
