% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nodes.R
\name{make_nodes}
\alias{make_nodes}
\title{Make a node df from the edges df}
\usage{
make_nodes(edgedf, groupers = NULL, typeorder = "werp")
}
\arguments{
\item{edgedf}{a tibble or dataframe of edges with \code{from} and \code{to} columns, as
well as \code{fromtype} and \code{totype}}

\item{groupers}{a character vector of grouping variables that might appear in
edgedf. Typically used to retain separate nodes for different planning
units or gauges}

\item{typeorder}{character vector or dataframe giving the order of the node
types in the network, or \code{werp} (the default). \code{werp} flags the use of the
default ranking}
}
\value{
a tibble with columns for \code{Name} of each individual node, \code{NodeType}
for the group of nodes of a given type, and \code{noderoder} specifying where
that nodetype is in the causal flow. May also include grouping column(s)
}
\description{
Takes a dataframe of edges and returns all the from and to nodes and their
nodetype, and attach a column indicating their position along the causal
sequence
}
