% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpys.R
\name{make_output_dir_R}
\alias{make_output_dir_R}
\title{Wrapper for make_output_dir}
\usage{
make_output_dir_R(...)
}
\arguments{
\item{...}{arguments for py function}
}
\value{
build the directory structure
}
\description{
Wrapper for make_output_dir
}
