% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpys.R
\name{paths_gauges_R}
\alias{paths_gauges_R}
\title{Wrapper for paths_gauges}
\usage{
paths_gauges_R(...)
}
\arguments{
\item{...}{arguments to the py function}
}
\value{
paths
}
\description{
Wrapper for paths_gauges
}
