% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hydrographs.R
\name{plot_hydrographs}
\alias{plot_hydrographs}
\title{Standard hydrograph plot}
\usage{
plot_hydrographs(
  outdf,
  y_col,
  scenariofilter = NULL,
  colors = "RColorBrewer::Dark2",
  scales = "fixed",
  transy = "identity",
  base_lev = NULL,
  comp_fun = NULL,
  ...
)
}
\arguments{
\item{y_col}{character, column name for what's plotted on the y-axis. Default 'flow', but will need to change if fed data with a different name}

\item{scenariofilter}{set of scenarios to plot, default \code{NULL} plots all of them}

\item{colors}{a named \code{colors} object or character vector giving a {paletteer} \code{palette} argument. Typically the former using \code{make_pal} to keep scenarios with consistent colours throughout, likely with a reference level.}

\item{scales}{facet scales, as in \code{?ggplot2::facet_wrap}. Default \code{scales = 'fixed'} holds them the same, most common change will be to \code{scales = 'free_y'} if gauges have very differen flows.}

\item{transy}{transformation for y axis as in \code{?ggplot2::scale_y_continuous}. Default \code{transy = 'identity'} just uses the data. Most common change likely \verb{transy = 'log10}}

\item{base_lev}{value to use as the base for comparison. Default NULL, no comparison. See \code{\link[=baseline_compare]{baseline_compare()}} and \code{\link[=create_base]{create_base()}} for options.}

\item{comp_fun}{function to use in comparison. Default NULL, no comparison. See \code{\link[=baseline_compare]{baseline_compare()}} and \code{\link[=create_base]{create_base()}} for options.}

\item{...}{}

\item{hydrolong}{hydrograph data in long format}

\item{gaugefilter}{set of gauges to plot, default \code{NULL} plots all of them}
}
\value{
ggplot object
}
\description{
Standard hydrograph plot
}
