% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcomes_bar.R
\name{plot_outcomes_bar}
\alias{plot_outcomes_bar}
\title{Plot outcomes as bar graph with scenario on x, possibly facetted and colored}
\usage{
plot_outcomes_bar(
  outdf,
  y_col,
  y_lab = y_col,
  facet_row = NULL,
  facet_col = NULL,
  facet_wrapper = NULL,
  scenariofilter = NULL,
  colors = "scico::oslo",
  sceneorder = NULL,
  scales = "fixed",
  transy = "identity",
  base_lev = NULL,
  comp_fun = NULL,
  ...
)
}
\arguments{
\item{outdf}{dataframe of outcomes, needs a \code{scenario} column}

\item{y_col}{character, name of column for the y-axis}

\item{y_lab}{character, default \code{y_col}, allows changing y-label}

\item{facet_row}{NULL (default) or character for facet row. Can be \code{'.'} to have one row and column-facets}

\item{facet_col}{NULL (default) or character for facet column. Can be \code{'.'} to have one column and row-facets}

\item{facet_wrapper}{NULL (default) or character for column to use for \code{facet_wrap}}

\item{scenariofilter}{set of scenarios to plot, default \code{NULL} plots all of
them}

\item{colors}{named \code{colors} object specifying the colors for the scenarios}

\item{sceneorder}{character or factor giving the order to present scenario levels}

\item{scales}{facet scales, as in \code{?ggplot2::facet_wrap}. Default \code{scales = 'fixed'} holds them the same, most common change will be to \code{scales = 'free_y'} if gauges have very differen flows.}

\item{transy}{transformation for y axis as in \code{?ggplot2::scale_y_continuous}. Default \code{transy = 'identity'} just uses the data. Most common change likely \verb{transy = 'log10}}

\item{base_lev}{value to use as the base for comparison. Default NULL, no
comparison. See \code{\link[=baseline_compare]{baseline_compare()}} and \code{\link[=create_base]{create_base()}} for options.}

\item{comp_fun}{function to use in comparison. Default NULL, no comparison.
See \code{\link[=baseline_compare]{baseline_compare()}} and \code{\link[=create_base]{create_base()}} for options.}

\item{...}{}
}
\value{
a ggplot bar graph with standardized style and color, possibly facetted
}
\description{
This is a simpler version of \code{plot_outcomes_stacked}
}
