% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{plot_prep}
\alias{plot_prep}
\title{Some standard data preparation for plotting}
\usage{
plot_prep(
  data,
  y_col,
  colors = "ggsci::default_igv",
  sceneorder = NULL,
  gaugefilter = NULL,
  scenariofilter = NULL,
  base_lev = NULL,
  comp_fun = NULL,
  ...
)
}
\arguments{
\item{data}{dataframe to prep}

\item{y_col}{character, column name for what's plotted on the y-axis. Default
'flow', but will need to change if fed data with a different name}

\item{colors}{a named \code{colors} object or character vector giving a
{paletteer} \code{palette} argument. Typically the former using \code{make_pal} to
keep scenarios with consistent colours throughout, likely with a reference
level.}

\item{gaugefilter}{set of gauges to plot, default \code{NULL} plots all of them}

\item{scenariofilter}{set of scenarios to plot, default \code{NULL} plots all of
them}

\item{base_lev}{value to use as the base for comparison. Default NULL, no
comparison. See \code{\link[=baseline_compare]{baseline_compare()}} and \code{\link[=create_base]{create_base()}} for options.}

\item{comp_fun}{function to use in comparison. Default NULL, no comparison.
See \code{\link[=baseline_compare]{baseline_compare()}} and \code{\link[=create_base]{create_base()}} for options.}

\item{...}{}
}
\value{
a list with prepped versions of \code{data}, \code{y_col}, \code{colors},
\code{gaugefilter}, \code{scenariofilter}, \code{base_lev}, \code{comp_fun}, \code{ylab_append} to
be used in plot calls
}
\description{
This is an attempt to pull a lot of copy-paste out of the top of plot
functions. It's likely to continue to evolve quite a bit as we see all the
different things we need to manage in the plot setup. One of the goals here
is to avoid making a million very similar datasets- doing it in functions
keeps those changes sandboxed
}
