% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_ewr_agg.R
\name{prep_ewr_agg}
\alias{prep_ewr_agg}
\title{Prepare ewr results from the EWR tool for use in aggregation}
\usage{
prep_ewr_agg(ewrpath, type = "summary", geopath, whichcrs = 4283, ...)
}
\arguments{
\item{ewrpath}{path to the EWR results output by the EWR tool (with toolkit
modifications to format)}

\item{type}{character, default \code{'summary'}, or \code{'annual'}.}

\item{geopath}{path to the file with gauge locations in lat/long (assumes BOM
currently)}

\item{whichcrs}{desired crs for the output}

\item{...}{passes \code{gaugefilter} and \code{scenariofilter} to \code{\link[=get_ewr_output]{get_ewr_output()}}.
Particularly useful for only processing some of the data at a time (e.g. in
parallel)}
}
\value{
an \code{sf} dataframe with the desired EWR outputs including scenario
names and gauge locations as a POINT geometry. Primary purpose is to
prepare data to send to \code{\link[=multi_aggregate]{multi_aggregate()}}
}
\description{
This reads in the EWR results given a path and \code{type} (summary or annual),
and pairs them to their geographic information
}
