% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpys.R
\name{prep_run_save_ewrs_R}
\alias{prep_run_save_ewrs_R}
\title{Wrapper for prep_run_save_ewrs}
\usage{
prep_run_save_ewrs_R(...)
}
\arguments{
\item{...}{All params for python fun}
}
\value{
EWR outputs if desired
}
\description{
Wrapper for prep_run_save_ewrs
}
