% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hydrographs.R
\name{plot_hydrographs}
\alias{plot_hydrographs}
\title{Standard hydrograph plot}
\usage{
plot_hydrographs(
  hydrolong,
  gaugefilter = NULL,
  scenariofilter = NULL,
  colors = "RColorBrewer::Dark2",
  scales = "fixed",
  transy = "identity"
)
}
\arguments{
\item{hydrolong}{hydrograph data in long format}

\item{gaugefilter}{set of gauges to plot, default \code{NULL} plots all of them}

\item{scenariofilter}{set of scenarios to plot, default \code{NULL} plots all of them}

\item{colors}{a named \code{colors} object or character vector giving a {paletteer} \code{palette} argument. Typically the former using \code{make_pal} to keep scenarios with consistent colours throughout, likely with a reference level.}

\item{scales}{facet scales, as in \code{?ggplot2::facet_wrap}. Default \code{scales = 'fixed'} holds them the same, most common change will be to \code{scales = 'free_y'} if gauges have very differen flows.}

\item{transy}{transformation for y axis as in \code{?ggplot2::scale_y_continuous}. Default \code{transy = 'identity'} just uses the data. Most common change likely \verb{transy = 'log10}}
}
\value{
ggplot object
}
\description{
Standard hydrograph plot
}
